feat: Extract styles in file
