+++ source/app/components/pages/dashboard/styles.ts
@@ -1,0 +1,14 @@
+export default {
+  gridContainer: {
+    bgcolor: 'background.default',
+    height: '100vh',
+    padding: '10px 20px'
+  },
+  textContainer: {
+    margin: '10px 0px'
+  },
+  divider: {
+    borderColor: 'white',
+    width: '100%'
+  }
+}
